Correlation Trade Bot
This script follows the price movement of the stimulus cryptocurrency and the price movement of the other cryptocurrency that correlates with it, when you enter the two correlated tickers in the fields shown.
After the stimulus sees the cryptocurrency movement, it gives a warning message if this movement has not yet occurred in the other cryptocurrency.
Of course, for this, you need to fill in the blanks in the code with the information of your discord channel. For this, you must create a discord bot and add it to a channel.
Developed by @sierra236 (Doğan Yiğit ŞAHİN) and @OgulcanYildiz (Oğulcan Yıldız).